+++ apps/designer-standalone/src/app/Designer/designer.tsx
@@ -1,6 +1,7 @@
 import { SettingsBox } from '../../components/settings_box';
 import type { RootState } from '../../state/store';
 import { HttpClient } from './httpClient';
+import { PseudoCommandBar } from './pseudoCommandBar';
 import {
   StandardConnectionService,
   StandardOperationManifestService,
@@ -8,6 +9,7 @@ import {
   StandardOAuthService,
   StandardGatewayService,
 } from '@microsoft/designer-client-services-logic-apps';
+import type { ContentType } from '@microsoft/designer-client-services-logic-apps';
 import { DesignerProvider, BJSWorkflowProvider, Designer } from '@microsoft/logic-apps-designer';
 import { ResourceIdentityType } from '@microsoft/utils-logic-apps';
 import { useEffect } from 'react';
@@ -67,14 +69,18 @@ const gatewayService = new StandardGatewayService({
 
 const workflowService = { getCallbackUrl: () => Promise.resolve({ method: 'POST', value: 'Dummy url' }) };
 
+const hostService = { fetchAndDisplayContent: (title: string, url: string, type: ContentType) => console.log(title, url, type) };
+
 export const DesignerWrapper = () => {
-  const { workflowDefinition, readOnly, monitoringView, darkMode, connections } = useSelector((state: RootState) => state.workflowLoader);
+  const { workflowDefinition, readOnly, monitoringView, darkMode, consumption, connections, runInstance } = useSelector(
+    (state: RootState) => state.workflowLoader
+  );
   const designerProviderProps = {
-    services: { connectionService, operationManifestService, searchService, oAuthService, gatewayService, workflowService },
+    services: { connectionService, operationManifestService, searchService, oAuthService, gatewayService, workflowService, hostService },
     readOnly,
     isMonitoringView: monitoringView,
     isDarkMode: darkMode,
-    isConsumption: false,
+    isConsumption: consumption,
   };
 
   useEffect(() => document.body.classList.add('is-standalone'), []);
@@ -88,8 +94,11 @@ export const DesignerWrapper = () => {
             workflow={{
               definition: workflowDefinition,
               connectionReferences: connections,
+              parameters: workflowDefinition.parameters,
             }}
+            runInstance={runInstance}
           >
+            <PseudoCommandBar />
             <Designer />
           </BJSWorkflowProvider>
         ) : null}