+++ libs/data-mapper/src/lib/core/state/DataMapSlice.ts
@@ -9,7 +9,7 @@ import type { SchemaExtended, SchemaNodeDictionary, SchemaNodeExtended } from '.
 import { SchemaNodeProperty, SchemaType } from '../../models';
 import type { ConnectionDictionary, InputConnection } from '../../models/Connection';
 import type { FunctionData, FunctionDictionary } from '../../models/Function';
-import { indexPseudoFunction } from '../../models/Function';
+import { indexPseudoFunction, directAccessPseudoFunctionKey } from '../../models/Function';
 import { findLast } from '../../utils/Array.Utils';
 import {
   bringInParentSourceNodesForRepeating,
@@ -367,56 +367,57 @@ export const dataMapSlice = createSlice({
       };
 
       addConnection(newState.dataMapConnections, action.payload);
-      const afterConnectionMade = { ...newState.dataMapConnections };
 
-      // Add any repeating parent nodes as well
+      // Add any repeating parent nodes as well (except for Direct Access's)
       // Get all the source nodes in case we have sources from multiple source chains
       const originalSourceNode = action.payload.source;
       let actualSources: SchemaNodeExtended[];
-      if (isFunctionData(originalSourceNode)) {
-        const sourceNodes = getConnectedSourceSchemaNodes(
-          [newState.dataMapConnections[action.payload.reactFlowSource]],
-          newState.dataMapConnections
-        );
-        actualSources = sourceNodes;
-      } else {
-        actualSources = [originalSourceNode];
-      }
 
-      // We'll only have one output node in this case
-      const originalTargetNode = action.payload.destination;
-      let actualTarget: SchemaNodeExtended[];
-      if (isFunctionData(originalTargetNode)) {
-        const targetNodes = getConnectedTargetSchemaNodes(
-          [newState.dataMapConnections[action.payload.reactFlowDestination]],
-          newState.dataMapConnections
-        );
-        actualTarget = targetNodes;
-      } else {
-        actualTarget = [originalTargetNode];
-      }
-
-      actualSources.forEach((sourceNode) => {
-        if (actualTarget.length > 0) {
-          addParentConnectionForRepeatingElementsNested(
-            sourceNode,
-            actualTarget[0],
-            newState.flattenedSourceSchema,
-            newState.flattenedTargetSchema,
+      if (!(isFunctionData(originalSourceNode) && originalSourceNode.key === directAccessPseudoFunctionKey)) {
+        if (isFunctionData(originalSourceNode)) {
+          const sourceNodes = getConnectedSourceSchemaNodes(
+            [newState.dataMapConnections[action.payload.reactFlowSource]],
             newState.dataMapConnections
           );
-
-          // If new parent connection has been made
-          if (JSON.stringify(newState.dataMapConnections) !== JSON.stringify(afterConnectionMade)) {
-            state.notificationData = { type: NotificationTypes.ArrayConnectionAdded };
-          }
-
-          // Bring in correct source nodes
-          // Loop through parent nodes connected to
-          const parentTargetNode = state.curDataMapOperation.currentTargetSchemaNode;
-          bringInParentSourceNodesForRepeating(parentTargetNode, newState);
+          actualSources = sourceNodes;
+        } else {
+          actualSources = [originalSourceNode];
         }
-      });
+
+        // We'll only have one output node in this case
+        const originalTargetNode = action.payload.destination;
+        let actualTarget: SchemaNodeExtended[];
+        if (isFunctionData(originalTargetNode)) {
+          const targetNodes = getConnectedTargetSchemaNodes(
+            [newState.dataMapConnections[action.payload.reactFlowDestination]],
+            newState.dataMapConnections
+          );
+          actualTarget = targetNodes;
+        } else {
+          actualTarget = [originalTargetNode];
+        }
+
+        actualSources.forEach((sourceNode) => {
+          if (actualTarget.length > 0) {
+            const wasNewArrayConnectionAdded = addParentConnectionForRepeatingElementsNested(
+              sourceNode,
+              actualTarget[0],
+              newState.flattenedSourceSchema,
+              newState.flattenedTargetSchema,
+              newState.dataMapConnections
+            );
+
+            if (wasNewArrayConnectionAdded) {
+              state.notificationData = { type: NotificationTypes.ArrayConnectionAdded };
+            }
+
+            // Bring in correct source nodes
+            // Loop through parent nodes connected to
+            const parentTargetNode = state.curDataMapOperation.currentTargetSchemaNode;
+            bringInParentSourceNodesForRepeating(parentTargetNode, newState);
+          }
+        });
+      }
 
       doDataMapOperation(state, newState);
     },
@@ -733,40 +734,44 @@ export const deleteNodeWithKey = (curDataMapState: DataMapState, reactFlowKey: s
   }
 
   // Item to be deleted is a connection
-  const sourceId = getSourceIdFromReactFlowConnectionId(reactFlowKey);
-  const sourceSchema = curDataMapState.curDataMapOperation.flattenedSourceSchema;
-  const canDelete = canDeleteConnection(sourceId, sourceSchema);
   const connections = { ...curDataMapState.curDataMapOperation.dataMapConnections };
 
-  if (canDelete) {
-    deleteConnectionFromConnections(
-      connections,
-      getSourceIdFromReactFlowConnectionId(reactFlowKey),
-      getDestinationIdFromReactFlowConnectionId(reactFlowKey)
-    );
-    const tempConn = connections[getSourceIdFromReactFlowConnectionId(reactFlowKey)];
-    const ids = getConnectedSourceSchemaNodes([tempConn], connections);
-    if (ids.length > 0) {
-      deleteParentRepeatingConnections(connections, addSourceReactFlowPrefix(ids[0].key));
+  /*
+    const sourceId = getSourceIdFromReactFlowConnectionId(reactFlowKey);
+    const sourceSchema = curDataMapState.curDataMapOperation.flattenedSourceSchema;
+    const canDelete = canDeleteConnection(sourceId, sourceSchema);
+
+    if (!canDelete) {
+      const sourceNode = sourceSchema[sourceId];
+      curDataMapState.notificationData = {
+        type: NotificationTypes.RepeatingConnectionCannotDelete,
+        msgParam: sourceNode.name,
+        autoHideDurationMs: errorNotificationAutoHideDuration,
+      };
     }
 
-    curDataMapState.notificationData = {
-      type: NotificationTypes.ConnectionDeleted,
-      autoHideDurationMs: deletedNotificationAutoHideDuration,
-    };
-  } else {
-    const sourceNode = sourceSchema[sourceId];
-    curDataMapState.notificationData = {
-      type: NotificationTypes.RepeatingConnectionCannotDelete,
-      msgParam: sourceNode.name,
-      autoHideDurationMs: errorNotificationAutoHideDuration,
-    };
+  */
+
+  deleteConnectionFromConnections(
+    connections,
+    getSourceIdFromReactFlowConnectionId(reactFlowKey),
+    getDestinationIdFromReactFlowConnectionId(reactFlowKey)
+  );
+  const tempConn = connections[getSourceIdFromReactFlowConnectionId(reactFlowKey)];
+  const ids = getConnectedSourceSchemaNodes([tempConn], connections);
+  if (ids.length > 0) {
+    deleteParentRepeatingConnections(connections, addSourceReactFlowPrefix(ids[0].key));
   }
 
   doDataMapOperation(curDataMapState, {
     ...curDataMapState.curDataMapOperation,
     dataMapConnections: { ...connections },
   });
+
+  curDataMapState.notificationData = {
+    type: NotificationTypes.ConnectionDeleted,
+    autoHideDurationMs: deletedNotificationAutoHideDuration,
+  };
 };
 
 export const addParentConnectionForRepeatingElements = (