+++ libs/data-mapper/src/lib/index.ts
@@ -2,3 +2,4 @@ export * from './ui/index';
 export * from './core/index';
 export * from './models/index';
 export * from './mapDefinitions/index';
+export * from './utils/Schema.Utils';