+++ libs/data-mapper/src/lib/models/Schema.ts
@@ -10,9 +10,6 @@ export interface SchemaNode {
   key: string;
   name: string;
   fullName: string;
-  parentKey?: string;
-  namespacePrefix?: string;
-  namespaceUri?: string;
   normalizedDataType: NormalizedDataType;
 
   /**
@@ -37,18 +34,21 @@ export enum SchemaNodeProperty {
   ComplexTypeSimpleContent = 'ComplexTypeSimpleContent',
   MaximumDepthLimit = 'MaximumDepthLimit',
   CyclicTypeReference = 'CyclicTypeReference',
+  JsonArray = 'JsonArray',
+  ArrayItem = 'ArrayItem',
+  AnyOf = 'AnyOf',
 }
 
 export enum NormalizedDataType {
-  ComplexType = 'ComplexType',
-  Integer = 'Integer',
-  Decimal = 'Decimal',
-  Number = 'Number',
+  Any = 'Any',
   Binary = 'Binary',
   Boolean = 'Bool',
-  String = 'String',
+  ComplexType = 'ComplexType',
   DateTime = 'DateTime',
-  Any = 'Any',
+  Decimal = 'Decimal',
+  Integer = 'Integer',
+  Number = 'Number',
+  String = 'String',
 }
 
 export interface SchemaExtended extends Schema {
@@ -60,7 +60,7 @@ export interface SchemaNodeExtended extends SchemaNode {
   nodeProperties: SchemaNodeProperty[];
   // Inclusive of the current node
   pathToRoot: PathItem[];
-  width?: number;
+  parentKey: string | undefined;
 }
 
 export interface PathItem {