+++ libs/designer/src/lib/core/index.ts
@@ -1,6 +1,7 @@
 export * from './DesignerProvider';
 export * from './BJSWorkflowProvider';
 export * from './ProviderWrappedContext';
+export { getReactQueryClient } from './ReactQueryProvider';
 export type { RootState, AppDispatch } from './store';
 export { store } from './store';
 export { discardAllChanges } from './state/workflow/workflowSlice';