+++ libs/designer/src/lib/core/utils/parameters/dynamicdata.ts
@@ -241,7 +241,12 @@ export function getDynamicOutputsFromSchema(schema: OpenAPIV2.SchemaObject, dyna
   const schemaProperties = new SchemaProcessor(processorOptions).getSchemaProperties(schema);
   const parameterLocation = equals(source, OutputSource.Outputs) ? OutputSource.Outputs : ParameterLocations.Body;
 
-  const outputParameters = schemaProperties.map((item) => ({ ...item, source, key: `${parameterLocation}.${item.key}` }));
+  let outputParameters = schemaProperties.map((item) => ({ ...item, source, key: `${parameterLocation}.${item.key}` }));
+
+  if (outputParameters.length > 1) {
+    outputParameters = outputParameters.filter((parameter) => parameter.key !== 'outputs.$');
+  }
+
   const outputs = map(outputParameters, OutputMapKey);
 
   return Object.keys(outputs).reduce(
@@ -371,7 +376,12 @@ function getManifestBasedInputParameters(
   operationDefinition: any
 ): InputParameter[] {
   let result: InputParameter[] = [];
-  const stepInputs = getInputsValueFromDefinitionForManifest(manifest.properties.inputsLocation ?? ['inputs'], operationDefinition);
+  const stepInputs = getInputsValueFromDefinitionForManifest(
+    manifest.properties.inputsLocation ?? ['inputs'],
+    manifest,
+    operationDefinition,
+    dynamicInputs
+  );
   const stepInputsAreNonEmptyObject = !isNullOrEmpty(stepInputs) && isObject(stepInputs);
 
   // Mark all of the known inputs as seen.
@@ -394,7 +404,7 @@ function getManifestBasedInputParameters(
     knownKeys.add(clonedInputParameter.key);
   }
 
-  if (stepInputs !== undefined) {
+  if (stepInputs !== undefined && !manifest.properties.inputsLocationSwapMap) {
     // load unknown inputs not in the schema by key.
     const resultParameters = map(result, 'key');
     loadUnknownManifestBasedParameters(keyPrefix, '', stepInputs, resultParameters, new Set<string>(), knownKeys);
@@ -532,7 +542,7 @@ function getObjectValue(key: string, currentObject: any): any {
 
   for (const keyPath of keyPaths) {
     const currentKey = tryConvertStringToExpression(decodePropertySegment(keyPath));
-    value = getPropertyValue(currentObject, currentKey);
+    value = getPropertyValue(currentValue, currentKey);
 
     if (value !== undefined) {
       currentValue = value;